managed deleting post without image

+++ back/controllers/post.js
@@ -276,8 +276,38 @@ exports.deletePost = (req, res, next) => {
         }
       );
       //deleting image and post
-      const filename = results[0].imageUrl.split("/images/")[1];
-      fs.unlink(`images/${filename}`, () => {
+      if (results[0].imageUrl) {
+        const filename = results[0].imageUrl.split("/images/")[1];
+        fs.unlink(`images/${filename}`, () => {
+          db.query(
+            "SELECT * FROM post WHERE id=?",
+            postId,
+            (error, results) => {
+              if (error) {
+                res.json({ error });
+              } else if (results == 0) {
+                return res.status(404).json({ error: "Post not found" });
+              } else {
+                db.query(
+                  "DELETE FROM post WHERE id =?",
+                  postId,
+                  (error, results) => {
+                    if (error) {
+                      res.json({ error });
+                    } else if (results == 0) {
+                      return res
+                        .status(404)
+                        .json({ error: "Post not deleted" });
+                    } else {
+                      res.status(200).json({ message: "post deleted" });
+                    }
+                  }
+                );
+              }
+            }
+          );
+        });
+      } else {
         db.query("SELECT * FROM post WHERE id=?", postId, (error, results) => {
           if (error) {
             res.json({ error });
@@ -299,7 +329,7 @@ exports.deletePost = (req, res, next) => {
             );
           }
         });
-      });
+      }
     }
   });
 };
